Added casts to remove warnings about type size changes losing information.
These should be done in a smarter fashion.  The full sizes should be
put into a temporary variable and then tested to see if they will fit
into the required space (uint8_t and int respectively).

+++ src/protocols/ab/cip.c
@@ -541,10 +541,10 @@ int cip_encode_tag_name(ab_tag_p tag,const char *name)
     /* word_count is in units of 16-bit integers, do not
      * count the word_count value itself.
      */
-    *word_count = ((dp - data)-1)/2;
+    *word_count = (uint8_t)((dp - data)-1)/2;
 
     /* store the size of the whole result */
-    tag->encoded_name_size = dp - data;
+    tag->encoded_name_size = (int)(dp - data);
 
     return 1;
 }
